clean up qmake and debian files
  * debian/mcompositor.aegis:
    Removed, it hasn't been effective and we use XRR to set the global
    alpha anyway.
  * debian/control:
    -- Depend on libgles2-sgx-img (dpkg-shlibdeps couldn't figure it out
       because of the odd name of the library).
    -- Make the -utils suggest python (because it contains a script).
    -- functional tests and l10n are Architecture: all
  * debian/rules:
    Make libmcompositor-users depend on our exact version to ensure
    synchronisation.
  * debuan/mcompositor-dev.install:
    Install development symlinks with the -dev package, not mcompositor.
  * meegotouch_config.pri:
    Removed obsolete/unnecessary/libmcompositor-only DEFINES.
  * mcompositor.pro:
    Fine-grained subdirectory dependencies and made the build unordered,
    which enables paralellization.
  * tests/tests.pro:
    Leave it up to the toplevel Makefile to build the appinterface test.
  * tests/focus-tracker/focus-tracker.pro
  * src/mcompositescene.cpp (MCompositeScene::drawItems):
    Build without warnings.

+++ src/mcompositescene.cpp
@@ -129,7 +129,7 @@ void MCompositeScene::drawItems(QPainter *painter, int numItems, QGraphicsItem *
         if (cw->isWindowTransitioning() || visible.intersects(r)
             || cw->type() == MCompositeWindowGroup::Type) {
             if (size >= 9)
-                to_paint.resize(to_paint.size()+1);
+                to_paint.resize((unsigned)to_paint.size()+1);
             to_paint[size++] = i;
             if (cw->propertyCache()->windowTypeAtom()
                                          == ATOM(_NET_WM_WINDOW_TYPE_DESKTOP))
